requirements.txt
baidu-aip==2.2.18.0
pyperclip==1.8.0
